requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
anyio=3.6.2=pyhd8ed1ab_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h5764c6d_3
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.2.0=pyh71513ae_0
babel=2.12.1=pyhd8ed1ab_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py310h06a4308_0
blas=1.0=mkl
bleach=6.0.0=pyhd8ed1ab_0
bottleneck=1.3.5=py310ha9d4c09_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=39.0.1=py310h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cython=0.29.34=py310heca2aa9_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=py310h06a4308_0
flit-core=3.8.0=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
gdown=4.7.1=pyhd8ed1ab_0
gensim=4.3.0=py310h1128e8f_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.4=py310h06a4308_0
importlib-metadata=6.4.1=pyha770c72_0
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.12.0=py310h06a4308_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.6=pyhd8ed1ab_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=py310h06a4308_0
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h5eee18b_1
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=8.1.0=py310h06a4308_0
jupyter_core=5.3.0=py310h06a4308_0
jupyter_server=1.23.6=pyhd8ed1ab_0
jupyterlab=3.5.3=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.22.1=pyhd8ed1ab_0
jupyterlab_widgets=3.0.7=pyhd8ed1ab_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib-inline=0.1.6=py310h06a4308_0
mistune=2.0.5=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=pypi_0
nbclassic=0.5.5=pyh8b2e9e2_0
nbclient=0.7.3=pyhd8ed1ab_0
nbconvert-core=7.3.1=pyhd8ed1ab_0
nbformat=5.8.0=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py310h06a4308_1
notebook=6.5.4=pyha770c72_0
notebook-shim=0.2.2=pyhd8ed1ab_0
numexpr=2.8.4=py310h8879344_0
numpy=1.24.2=py310h8deb116_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h0b41bf4_0
packaging=23.0=py310h06a4308_0
pandas=1.5.3=py310h1128e8f_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=2.5.2=py310h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prometheus_client=0.16.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=py310h06a4308_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.3.0=py310_torch_2.0.0_cu118
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pyrsistent=0.19.3=py310h1fa729e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.3=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=2.0.0=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cluster=1.6.1=py310_torch_2.0.0_cu118
pytorch-cuda=11.8=h7e8668a_3
pytorch-mutex=1.0=cuda
pytorch-scatter=2.1.1=py310_torch_2.0.0_cu118
pytorch-sparse=0.6.17=py310_torch_2.0.0_cu118
pytorch-spline-conv=1.2.2=py310_torch_2.0.0_cu118
pytz=2022.7=py310h06a4308_0
pyzmq=23.2.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_1
scikit-learn=1.2.2=py310h6a678d5_0
scipy=1.10.1=py310hd5efca6_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=66.0.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py310h06a4308_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.4=py310h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py310h06a4308_0
tabulate=0.9.0=pyhd8ed1ab_1
terminado=0.17.1=pyh41d4057_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.0.0=py310_cu118
torchinfo=1.7.2=pyhd8ed1ab_0
torchtriton=2.0.0=py310
torchvision=0.15.0=py310_cu118
tornado=6.2=py310h5eee18b_0
tqdm=4.65.0=py310h2f386ee_0
traitlets=5.7.1=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
websocket-client=1.5.1=pyhd8ed1ab_0
wheel=0.38.4=py310h06a4308_0
widgetsnbextension=4.0.7=pyhd8ed1ab_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
